Move component into subcomponent folder.
