+++ app/components/ScenariosManager/ScenarioHelper/ScenarioHelper.module.css
@@ -1,6 +1,6 @@
 .helper {
-    --spacing-sm: 0.75rem;    /* 12px */
-    --spacing-md: 1.5rem;     /* 24px */
+    --spacing-sm: 0.75rem;    
+    --spacing-md: 1.5rem;     
     --border-color: #e8e8e8;
 }
 
@@ -21,7 +21,7 @@
 .scenarioImage {
     width: 100%;
     height: auto;
-    max-width: 62.5rem;  /* 1000px */
+    max-width: 62.5rem;
     margin-bottom: var(--spacing-md);
 }
